more reorganization, start to flatten things out a bit more
