--- .dockerignore
@@ -1,11 +1,0 @@
-# CI configuration
-.circleci
-
-# Build artifacts
-build
-
-# Left over executables
-synse
-
-# Autocomplete files
-bash_autocomplete
