+++ pkg/commands/hosts/active.go
@@ -3,20 +3,26 @@ package hosts
 import (
 	"github.com/urfave/cli"
 	"github.com/vapor-ware/synse-cli/pkg/config"
-	"github.com/vapor-ware/synse-cli/pkg/flags"
-	"github.com/vapor-ware/synse-cli/pkg/utils"
 	"github.com/vapor-ware/synse-cli/pkg/formatters"
+	"github.com/vapor-ware/synse-cli/pkg/utils"
 )
 
 // hostsActiveCommand is the CLI sub-command for getting the active host.
 var hostsActiveCommand = cli.Command{
 	Name:  "active",
 	Usage: "Display information for the active host",
-	Flags: []cli.Flag{
-		flags.OutputFlag,
-	},
+
 	Action: func(c *cli.Context) error {
 		return utils.CmdHandler(cmdActive(c))
+	},
+
+	Flags: []cli.Flag{
+		// --output, -o flag specifies the output format (YAML, JSON) for the command
+		cli.StringFlag{
+			Name:  "output, o",
+			Value: "yaml",
+			Usage: "set the output format of the command",
+		},
 	},
 }
