+++ pkg/commands/plugin/cmd.go
@@ -7,7 +7,6 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/urfave/cli"
-	"github.com/vapor-ware/synse-cli/pkg/flags"
 	"github.com/vapor-ware/synse-server-grpc/go"
 	"google.golang.org/grpc"
 )
@@ -16,15 +15,25 @@ import (
 var PluginCommand = cli.Command{
 	Name:  "plugin",
 	Usage: "Interact with Synse Plugins",
-	Flags: []cli.Flag{
-		flags.HostFlag,
-		flags.UnixFlag,
-	},
+
 	Subcommands: []cli.Command{
 		pluginMetainfoCommand,
 		pluginReadCommand,
 		pluginTransactionCommand,
 		pluginWriteCommand,
+	},
+
+	Flags: []cli.Flag{
+		// --tcp, -t flag specifies the TCP address for the plugin to interface with
+		cli.StringFlag{
+			Name:  "tcp, t",
+			Usage: "set the hostname/ip[:port] for a plugin",
+		},
+		// --unix, -u flag specifies the UNIX socket for the plugin to interface with
+		cli.StringFlag{
+			Name:  "unix, u",
+			Usage: "set the unix socket path for a plugin",
+		},
 	},
 }
