+++ pkg/commands/server/version.go
@@ -3,7 +3,6 @@ package server
 import (
 	"github.com/urfave/cli"
 	"github.com/vapor-ware/synse-cli/pkg/client"
-	"github.com/vapor-ware/synse-cli/pkg/flags"
 	"github.com/vapor-ware/synse-cli/pkg/formatters"
 	"github.com/vapor-ware/synse-cli/pkg/scheme"
 	"github.com/vapor-ware/synse-cli/pkg/utils"
@@ -32,11 +31,18 @@ var VersionCommand = cli.Command{
 	Description: versionCmdDesc,
 	Category:    SynseActionsCategory,
 	ArgsUsage:   utils.NoArgs,
-	Flags: []cli.Flag{
-		flags.OutputFlag,
-	},
+
 	Action: func(c *cli.Context) error {
 		return utils.CmdHandler(cmdVersion(c))
+	},
+
+	Flags: []cli.Flag{
+		// --output, -o flag specifies the output format (YAML, JSON) for the command
+		cli.StringFlag{
+			Name:  "output, o",
+			Value: "yaml",
+			Usage: "set the output format of the command",
+		},
 	},
 }
